third commit

+++ src/blog/api/views.py
@@ -1,8 +1,15 @@
+from re import search
+from django.contrib.auth.models import User
 from rest_framework import status
+from rest_framework import pagination
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
-
+from rest_framework.decorators import api_view, authentication_classes, permission_classes
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.pagination import PageNumberPagination
+from rest_framework.generics import ListAPIView
+from rest_framework.filters import SearchFilter, OrderingFilter
 from account.models import Account
+from rest_framework.authentication import TokenAuthentication
 from blog.models import BlogPost
 from blog.api.serializers import BlogPostSerializer
 
@@ -13,6 +20,7 @@ UPDATE_SUCCESS = 'updated'
 CREATE_SUCCESS = 'created'
 
 @api_view(['GET', ])
+@permission_classes((IsAuthenticated))
 def api_detail_blog_view(request, slug):
 
 	try:
@@ -26,12 +34,19 @@ def api_detail_blog_view(request, slug):
 
 
 @api_view(['PUT',])
+@permission_classes((IsAuthenticated))
 def api_update_blog_view(request, slug):
 
 	try:
 		blog_post = BlogPost.objects.get(slug=slug)
 	except BlogPost.DoesNotExist:
 		return Response(status=status.HTTP_404_NOT_FOUND)
+
+	user = request.user
+	if blog_post.author != user:
+		return Response({'response': "You are not authorized to change the content"})
+     
+
 
 	if request.method == 'PUT':
 		serializer = BlogPostSerializer(blog_post, data=request.data)
@@ -44,13 +59,19 @@ def api_update_blog_view(request, slug):
 
 
 @api_view(['DELETE',])
+@permission_classes((IsAuthenticated))
+
 def api_delete_blog_view(request, slug):
 
 	try:
 		blog_post = BlogPost.objects.get(slug=slug)
 	except BlogPost.DoesNotExist:
 		return Response(status=status.HTTP_404_NOT_FOUND)
-
+	
+	user = request.user
+	if blog_post.author != user:
+		return Response({'response': "You are not authorized to change the content"})
+	
 	if request.method == 'DELETE':
 		operation = blog_post.delete()
 		data = {}
@@ -60,10 +81,12 @@ def api_delete_blog_view(request, slug):
 
 
 @api_view(['POST'])
+@permission_classes((IsAuthenticated))
+
 def api_create_blog_view(request):
 
-	account = Account.objects.get(pk=1)
-
+	# account = Account.objects.get(pk=1)
+	account =  request.user
 	blog_post = BlogPost(author=account)
 
 	if request.method == 'POST':
@@ -75,5 +98,11 @@ def api_create_blog_view(request):
 		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-
-
+class ApiBlogListView(ListAPIView):
+    queryset = BlogPost.objects.all()
+    serializer = BlogPostSerializer
+    authentication_classes= (TokenAuthentication,)
+    permission_classes = (IsAuthenticated,)
+    pagination_class= PageNumberPagination
+    filter_back = (SearchFilter,OrderingFilter)	
+    search_fields = ('title' , 'body','author__username')
